+++ CHANGES.rst
@@ -3,6 +3,8 @@ Version 0.2.0
 
 Release TBD
 
+- Renamed to Doozer-S3
+
 Version 0.1.0
 -------------